+++ rnd/autogpt_builder/src/lib/autogpt-server-api/client.ts
@@ -1,5 +1,12 @@
-import { XYPosition } from "reactflow";
-import { ObjectSchema } from "./types";
+import {
+  Block,
+  Graph,
+  GraphCreatable,
+  GraphUpdateable,
+  GraphMeta,
+  GraphExecuteResponse,
+  NodeExecutionResult,
+} from "./types"
 
 export default class AutoGPTServerAPI {
   private baseUrl: string;
@@ -196,96 +203,12 @@ export default class AutoGPTServerAPI {
   }
 }
 
-/* Mirror of autogpt_server/data/block.py:Block */
-export type Block = {
-  id: string;
-  name: string;
-  description: string;
-  inputSchema: ObjectSchema;
-  outputSchema: ObjectSchema;
-};
 
-/* Mirror of autogpt_server/data/graph.py:Node */
-export type Node = {
-  id: string;
-  block_id: string;
-  input_default: { [key: string]: any };
-  input_nodes: Array<{ name: string, node_id: string }>;
-  output_nodes: Array<{ name: string, node_id: string }>;
-  metadata: {
-    position: XYPosition;
-    [key: string]: any;
-  };
-};
+/* *** UTILITY TYPES *** */
 
-/* Mirror of autogpt_server/data/graph.py:Link */
-export type Link = {
-  id: string;
-  source_id: string;
-  sink_id: string;
-  source_name: string;
-  sink_name: string;
-}
-
-export type LinkCreatable = Omit<Link, "id"> & {
-  id?: string;
-}
-
-/* Mirror of autogpt_server/data/graph.py:GraphMeta */
-export type GraphMeta = {
-  id: string;
-  version: number;
-  is_active: boolean;
-  is_template: boolean;
-  name: string;
-  description: string;
-}
-
-/* Mirror of autogpt_server/data/graph.py:Graph */
-export type Graph = GraphMeta & {
-  nodes: Array<Node>;
-  links: Array<Link>;
-};
-
-export type GraphUpdateable = Omit<
-  Graph,
-  "version" | "is_active" | "is_template" | "links"
-> & {
-  version?: number;
-  is_active?: boolean;
-  is_template?: boolean;
-  links: Array<LinkCreatable>;
-}
-
-export type GraphCreatable = Omit<GraphUpdateable, "id"> & { id?: string }
-
-export type GraphCreateRequestBody = {
+type GraphCreateRequestBody = {
   template_id: string;
   template_version: number;
 } | {
   graph: GraphCreatable;
 }
-
-/* Derived from autogpt_server/executor/manager.py:ExecutionManager.add_execution */
-export type GraphExecuteResponse = {
-  /* ID of the initiated run */
-  id: string;
-  /* List of node executions */
-  executions: Array<{ id: string, node_id: string }>;
-};
-
-/* Mirror of autogpt_server/data/execution.py:ExecutionResult */
-export type NodeExecutionResult = {
-  graph_exec_id: string;
-  node_exec_id: string;
-  graph_id: string;
-  graph_version: number;
-  node_id: string;
-  status: 'INCOMPLETE' | 'QUEUED' | 'RUNNING' | 'COMPLETED' | 'FAILED';
-  input_data: { [key: string]: any };
-  output_data: { [key: string]: Array<any> };
-  add_time: Date;
-  queue_time?: Date;
-  start_time?: Date;
-  end_time?: Date;
-};
