+++ rnd/autogpt_builder/src/lib/autogpt-server-api/index.ts
@@ -1,0 +1,4 @@
+import AutoGPTServerAPI from "./client";
+
+export default AutoGPTServerAPI;
+export * from "./types";
